Rename application integration tests project from Tests.Integration to Integration.Tests postfix.
