+++ Digdir.Domain.Dialogporten.sln
@@ -23,8 +23,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Digdir.Library.Entity.Abstr
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{CADB8189-4AA1-4732-844A-C41DBF3EC8B7}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Digdir.Domain.Dialogporten.Application.Tests.Integration", "tests\Digdir.Domain.Dialogporten.Application.Tests.Integration\Digdir.Domain.Dialogporten.Application.Tests.Integration.csproj", "{A8C5A8ED-3A4A-43F9-A083-6F20FCEBB773}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Digdir.Domain.Dialogporten.Infrastructure.Unit.Tests", "tests\Digdir.Domain.Dialogporten.Infrastructure.Unit.Tests\Digdir.Domain.Dialogporten.Infrastructure.Unit.Tests.csproj", "{38A843CE-EE54-4476-ADA1-D8084C3901A9}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{33E486B1-3279-4645-B536-AB640D560C45}"
@@ -36,6 +34,8 @@ EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Digdir.Domain.Dialogporten.ChangeDataCapture", "src\Digdir.Domain.Dialogporten.ChangeDataCapture\Digdir.Domain.Dialogporten.ChangeDataCapture.csproj", "{9420D9DA-6C2F-4664-B361-FA5E2E1FB6EB}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Digdir.Domain.Dialogporten.Service", "src\Digdir.Domain.Dialogporten.Service\Digdir.Domain.Dialogporten.Service.csproj", "{45A63837-99BC-469C-A2DE-FEE6975516C3}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Digdir.Domain.Dialogporten.Application.Integration.Tests", "tests\Digdir.Domain.Dialogporten.Application.Integration.Tests\Digdir.Domain.Dialogporten.Application.Integration.Tests.csproj", "{A1CC5E4E-020C-4ABA-BD66-48819259BDB3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -67,10 +67,6 @@ Global
 		{E7288616-9358-47C2-8650-EA019AB9D16F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E7288616-9358-47C2-8650-EA019AB9D16F}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E7288616-9358-47C2-8650-EA019AB9D16F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A8C5A8ED-3A4A-43F9-A083-6F20FCEBB773}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A8C5A8ED-3A4A-43F9-A083-6F20FCEBB773}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A8C5A8ED-3A4A-43F9-A083-6F20FCEBB773}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A8C5A8ED-3A4A-43F9-A083-6F20FCEBB773}.Release|Any CPU.Build.0 = Release|Any CPU
 		{38A843CE-EE54-4476-ADA1-D8084C3901A9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{38A843CE-EE54-4476-ADA1-D8084C3901A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{38A843CE-EE54-4476-ADA1-D8084C3901A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -83,6 +79,10 @@ Global
 		{45A63837-99BC-469C-A2DE-FEE6975516C3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{45A63837-99BC-469C-A2DE-FEE6975516C3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{45A63837-99BC-469C-A2DE-FEE6975516C3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A1CC5E4E-020C-4ABA-BD66-48819259BDB3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A1CC5E4E-020C-4ABA-BD66-48819259BDB3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A1CC5E4E-020C-4ABA-BD66-48819259BDB3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A1CC5E4E-020C-4ABA-BD66-48819259BDB3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -96,10 +96,10 @@ Global
 		{CC70CBB9-D57A-4095-8786-956012537B1C} = {096E9B69-6783-4446-A895-0B6D7729A0D9}
 		{0EE4AC8C-BDB9-474A-821D-23FB95CB54A7} = {CC70CBB9-D57A-4095-8786-956012537B1C}
 		{E7288616-9358-47C2-8650-EA019AB9D16F} = {CC70CBB9-D57A-4095-8786-956012537B1C}
-		{A8C5A8ED-3A4A-43F9-A083-6F20FCEBB773} = {CADB8189-4AA1-4732-844A-C41DBF3EC8B7}
 		{38A843CE-EE54-4476-ADA1-D8084C3901A9} = {CADB8189-4AA1-4732-844A-C41DBF3EC8B7}
 		{9420D9DA-6C2F-4664-B361-FA5E2E1FB6EB} = {320B47A0-5EB8-4B6E-8C84-90633A1849CA}
 		{45A63837-99BC-469C-A2DE-FEE6975516C3} = {320B47A0-5EB8-4B6E-8C84-90633A1849CA}
+		{A1CC5E4E-020C-4ABA-BD66-48819259BDB3} = {CADB8189-4AA1-4732-844A-C41DBF3EC8B7}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {B2FE67FF-7622-4AFB-AD8E-961B6A39D888}
