+++ tests/Digdir.Domain.Dialogporten.Application.Integration.Tests/Features/V1/Dialogs/Commands/UpdateDialogTests.cs
@@ -1,0 +1,15 @@
+﻿using Digdir.Domain.Dialogporten.Application.Integration.Tests.Common;
+using Digdir.Domain.Dialogporten.Application.Integration.Tests.Features.V1.Dialogs;
+using Xunit;
+
+namespace Digdir.Domain.Dialogporten.Application.Integration.Tests.Features.V1.Dialogs.Commands;
+
+[Collection(nameof(DialogCqrsCollectionFixture))]
+public class UpdateDialogTests : ApplicationCollectionFixture
+{
+    public UpdateDialogTests(DialogApplication application) : base(application)
+    {
+    }
+    // TODO: Add tests
+}
+
