--- tests/Digdir.Domain.Dialogporten.Application.Tests.Integration/Features/V1/Dialogs/Commands/UpdateDialogTests.cs
@@ -1,14 +1,0 @@
-﻿using Digdir.Domain.Dialogporten.Application.Tests.Integration.Common;
-using Xunit;
-
-namespace Digdir.Domain.Dialogporten.Application.Tests.Integration.Features.V1.Dialogs.Commands;
-
-[Collection(nameof(DialogCqrsCollectionFixture))]
-public class UpdateDialogTests : ApplicationCollectionFixture
-{
-    public UpdateDialogTests(DialogApplication application) : base(application)
-    {
-    }
-    // TODO: Add tests
-}
-
