--- tests/Digdir.Domain.Dialogporten.Application.Tests.Integration/Features/V1/Dialogs/Queries/ListDialogTests.cs
@@ -1,13 +1,0 @@
-﻿using Digdir.Domain.Dialogporten.Application.Tests.Integration.Common;
-using Xunit;
-
-namespace Digdir.Domain.Dialogporten.Application.Tests.Integration.Features.V1.Dialogs.Queries;
-
-[Collection(nameof(DialogCqrsCollectionFixture))]
-public class ListDialogTests : ApplicationCollectionFixture
-{
-    public ListDialogTests(DialogApplication application) : base(application)
-    {
-    }
-    // TODO: Add tests
-}
